Refactor: 도메인별 패키지 구조 정리
