+++ src/main/java/elice/aishortform/image/entity/ImageEntity.java
@@ -1,4 +1,4 @@
-package elice.aishortform.entity;
+package elice.aishortform.image.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
