+++ src/main/java/elice/aishortform/summary/controller/SummarizeApiDocs.java
@@ -1,7 +1,7 @@
-package elice.aishortform.controller;
+package elice.aishortform.summary.controller;
 
-import elice.aishortform.dto.SummarizeRequest;
-import elice.aishortform.dto.SummarizeResponse;
+import elice.aishortform.summary.dto.SummarizeRequest;
+import elice.aishortform.summary.dto.SummarizeResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
