+++ src/main/java/elice/aishortform/summary/dto/SummarizeRequest.java
@@ -1,4 +1,4 @@
-package elice.aishortform.dto;
+package elice.aishortform.summary.dto;
 
 public record SummarizeRequest(String url, String platform) {
 }
