+++ src/main/java/elice/aishortform/summary/repository/SummaryRepository.java
@@ -1,6 +1,6 @@
-package elice.aishortform.repository;
+package elice.aishortform.summary.repository;
 
-import elice.aishortform.entity.Summary;
+import elice.aishortform.summary.entity.Summary;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
